Remove bounding box check.
