+++ celeri/celeri_util.py
@@ -8,3 +8,9 @@ def sph2cart(lon, lat, radius):
     y = radius * np.cos(lat_rad) * np.sin(lon_rad)
     z = radius * np.sin(lat_rad)
     return x, y, z
+
+def cart2sph(x, y, z):
+    azimuth = np.arctan2(y, x)
+    elevation = np.arctan2(z, np.sqrt(x ** 2 + y ** 2))
+    r = np.sqrt(x ** 2 + y ** 2 + z ** 2)
+    return azimuth, elevation, r
